Fixed up some parts of the sleep cycle and the stat values for the vital stats. Also Changed weather from a direct map to an object containing a map for future simulation ease.

+++ Weather.cpp
@@ -9,7 +9,7 @@ void Weather::setWeatherMapSize(int size)
   weatherMap.setSize(size);
 }
 
-Map<double>* Weather::getWeatherMap()
+Map<int>* Weather::getWeatherMap()
 {
   return &weatherMap;
 }
